+++ research/tests/runtests.jl
@@ -1,5 +1,9 @@
-using AdvancedHMC
 using ReTest
 
-include("AdvancedHMCResearchTests.jl");
-AdvancedHMCResearchTests.runtests()
+# include the source code for experimental HMC
+include("../src/relativistic_hmc.jl")
+
+# include the tests for experimental HMC
+include("relativistic_hmc.jl")
+
+retest(; dry=false, verbose=Inf)
