+++ src/trajectory.jl
@@ -141,8 +141,9 @@ $(TYPEDEF)
 Slice sampler for the starting single leaf tree.
 Slice variable is initialized.
 """
-SliceTS(rng::AbstractRNG, z0::PhasePoint) =
+function SliceTS(rng::AbstractRNG, z0::PhasePoint)
     SliceTS(z0, neg_energy(z0) - Random.randexp(rng), 1)
+end
 
 """
 $(TYPEDEF)
@@ -292,7 +293,7 @@ function transition(
             hamiltonian_energy=H,
             hamiltonian_energy_error=H - H0,
             # check numerical error in proposed phase point. 
-            numerical_error=!all(isfinite, H′),
+            numerical_error=(!all(isfinite, H′)),
         ),
         stat(τ.integrator),
     )
@@ -727,7 +728,7 @@ function transition(
         (
             n_steps=tree.nα,
             is_accept=true,
-            acceptance_rate=tree.sum_α / tree.nα,
+            acceptance_rate=(tree.sum_α / tree.nα),
             log_density=zcand.ℓπ.value,
             hamiltonian_energy=H,
             hamiltonian_energy_error=H - H0,
